Relates to #14 The Command class paves way for a new way of developing smarter, class based command bindings with Pyttman.
It resembles the API of developing web apps with other frameworks by
defining a response method and optional field overloads for defining
customized behavior.

The InputStringParser inner class is such an improvement. It allows
the developer to define a class by ready-made Parser classes
(or subclassing and writing a customized one), to allow for a high-level
API to extract and identify strings of interest in a message.