Relates to #14
Identifier class

The Identifier class is an optional argument to a Parser subclass inside
the inner InputStringParser class used in Command subclasses.

It offers a simple interface for developers to find and extract strings
in messages, for different purposes. Many times, when trying to extract
strings of value such as in a question with metadata in it, having
prefix and suffix words to look for or after is not enough.
Sometimes the pattern of the value itself is more valuable.

The Identifier class aims to provide an intuitive and simple API for
these scenarios. By providing an Identifier class to a field inside a
InputStringParser class, the InputStringParser will use that Identifier
to fill that field.

Identifier classes are available to find phone numbers, integers, and
can be easily subclassed by writing a regex pattern to identify exactly
what the developer is looking for in that very command.