Relates to #14
Parser class

The Parser and its subclasses are a plugin to further the high-level
API experience when developing apps with Pyttman.

The need to extract data in a message is common and central in developing
virtual assistants and chatbots. The Parser classes provide an easy
API for this purpose, specialized in whichever type of data that the
subclass of Parser has defined.

The ValueParser is the most flexible one, and identifies values in the
message based on the prefixes, suffixes and/or the identifier arguments
which is is provided. Together, these parameters can create a very
loose or very granular control over which strings are identifies as
values for a given command.

In some cases, it is considered safe enough to select the first or
last element in the message, as the value sought for. The PositionalParser
is the way to go here, as you can define which element the value is parsed
from in the message with a simple field.