Improves parse_message method in Parser
Adds functionality to increase filtration if an identifier -and-
suffix or prefix tuples are used together.

As more data is provided, the finer the filtration. Prefixes, suffixes
and the identifier class will work together or alone, depending on
the configuration. When identifier is used with prefix and/or suffix
tuples, these will add conditions for the identifier only to be valid
if the identifier string is actually before suffixes and after prefixes
as they are defined.