Imrovement to the Message object
The Message object now offers a method, and a parmeter on as_list and
as_str methods called 'sanitized' which calls the 'sanitize' method.

The 'sanitize()' method in the Message class deletes non-alpha characters
but leaves numbers in each substring in the .content property.
This is useful for scenarios where the actual original content is
meant to be kept but the sanitized version is needed for parsing
commands.