Pyttman DiscordClient class
The DiscordClient class is implemented in the new Client API from Pyttman
as a way to allow Pyttman developers have a no-code experience when
developing an app and then launching it on Discord.

Without a single line of code, this class can be plugged in to a
Pyttman application and used to host a chatbot on the Discord platform.

The DiscordClient class inherits both from the BaseClient interface for
Pyttman internal compatibility, and also extends the discord.Client
class.

The discord.Client class on_message is overridden, to allow integration
with the Pyttman MessageRouter API for routing messages across Features
and Command classes.

On top of the discord.Client class, it also allows for users to set
'message_endswith' and/or 'message_startswith' attributes to increase
conditions on which messages reaches the Client and thus the Pyttman
app stack as a whole.

This is to omitt the typical if statement: `if not message.content.pop().startswith('!'): pass`
and alike.