DiscordMessage class
The DiscordMessage class is a Message container, implementing the
MessageMixin interface.

It combines the functionality of the discord.Message, and the
Pyttman.MessageMixin, to comply with the constraints set from both
API's while allowing seamless integration with both.

This makes the development experience seamless when using the
DiscordClient, and developing a Pyttman app to interface with the
DiscordClient class.