Class MessageMixin
The MessageMixin class is an implementable interface to allow for
clients of various platforms to interface with the internals of an app
developed in Pyttman. Behaviors such as the message content, the client
attribute (relative to which client sent the message) and all other
fields and methods associated with a Message-like class from an SDK or
API, can all be seamlessly integrated with a Pyttman app using the
MessageMixin baseclass.