Class Reply
the Reply class is designed to be returned from a Command.respond()
method in Pyttman apps. Much like the HttpResponse / Response classes
in Django, the Reply class formats the strings or objects the user may
provide to it, and display it properly in the client side.
It behaves much like a Message but also provides ability to extend it
for future needs.