helper function: _generate_error_entry
_generate_error_entry is an internal function for situations where an
internal error has occurred in the application or in Pyttman itself.

The user will normally in these situations be unaware that a problem has
occurred, and may not get an answer at all. This function will return
a Reply object with a phrase and a UUID for later reference.

The method also creates a log entry with more details about the exception
and mentions this UUID in the log entry for quick lookup in the log file.

The contents of the message is editable in settings.py for the app.