The InputString API :: ValueParser._identify_value method
the ValueParser class is one of the Parser classes which specializes
in identifying strings based on a set of conditions.

In cooperation with an optional set of prefixes, suffixes and/or an
Identifier class, it's very easy to fine tune the extraction of information
in a string of natural language.