function bootstrap_environment
The figurative 'ignition' of a Pyttman app.
This method imports the settings module of an app located
in the same directory as where the command is executed in.

Settings are parsed and fed to the logger, the clients, the features
and various other internal components of the Pyttman framework runtime.

It instantiates Runner classes which contextualize Client objects,
ready to be run in parallel.