Solves an issue with the use of an Identifier class inside a ValueParser, and the 'span' property.
The algorithm is rewritten to allow for further iteration of the
message content only if the Identifier is satisfied with the next
value in line - or iteration stops. This solves the issue where
Identifier classes would go beyond their scope and identify something
which is not considered valid according to the Identifier used.