Introducing the ReplyStream class
The ReplyStream class offers a simple interface for app developers
to return a collection of Reply objects from a Command in their app.

The Client will send each message in the ReplyStream as an individual
message. It's useful for situations where the response for the app can
be a list of items, where platforms have message length restrictions
or are otherwise cumbersome to format as one single big Reply.

The ReplyStream is a subclass of Queue and typecasts every element
output from the 'get' method as a Reply object.