Entity class
Class Entity

The Entity class is a lightweight class to represent an entity
in a message object.

It holds both its value and the index of position inside the message
in which it occurred from, for parsers to relate to where it is located
in the determination of whether another Entity class is actually the same
word in the message or not.