new class: Settings class The Settings class introduces an improved way of accessing Settings in Pyttman apps.
As the settings.py module was previously imported, it introduced a lot
of name space flooding with references never used, as well as function
calls and imports done inside the module.

The Settings class removes this bloat and also allows for multiple
Settings configs for later releases using DEBUG configurations.