Fixes #35
This commit brings a vast improvement to how the EntityParser uses its
subsequent parsers to traverse over the message and find the most likely
correctly matching entities, while also omitting words which have
delimiting properties due to their use in 'lead/trail/exclude' tuples
of the Parser class.

Messages are initially traversed left-to-right when identifying Entities
in Message object. However, a further right-to-left traverse (message
content parsed in reverse) after the fact, improves the ability for
the EntityParser to use set subtraction to truncate strings (words)
found to occur in more than one entity, later in the message.