Added functionality to prohibit the use of 'pyttman' module outside of the 'pyttman' cli
Users shouldn't import pyttman directly to access 'settings' variables.
At the time of import, the settings are not configured and pyttman.settings
were previously None.

This was ambiguous since it would only show when users tried to access
variables on the settings object during runtime.

this improvement prevents users from trying to access pyttman.settings
when outside of an app environment.