Refactored the 'Runner' component in Pyttman apps.
The Runner class was previously the highest level of abstraction for applications developed with Pyttman, but were not exposed to the app developer.
This changes that; now developers can import 'app' from pyttman, e.g. 'from pyttman import app' and access a module-level singleton instance of their
app.

This allows them to both inspect the state of their app, access its client instance towards the platform of choice; and also opens the door for further development down the road.
The first new component this refactor brings, is the 'hooks' attribute on the app, which can be used to decorate functions and store them as lifecycle hooks.